+++ app/authorization/service/service_test.go
@@ -1,4 +1,4 @@
-/* Copyright © 2020. Financelime, https://financelime.com. All rights reserved.
+/* Copyright © 2021. Financelime, https://financelime.com. All rights reserved.
    Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
    License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
 
@@ -10,10 +10,10 @@ import (
 	"github.com/dmalix/financelime-authorization/app/authorization/model"
 	"github.com/dmalix/financelime-authorization/app/authorization/repository"
 	"github.com/dmalix/financelime-authorization/config"
-	"github.com/dmalix/financelime-authorization/packages/cryptographer"
-	"github.com/dmalix/financelime-authorization/packages/email"
-	"github.com/dmalix/financelime-authorization/packages/jwt"
-	"github.com/dmalix/financelime-authorization/packages/middleware"
+	"github.com/dmalix/jwt"
+	"github.com/dmalix/middleware"
+	"github.com/dmalix/secretdata"
+	"github.com/dmalix/sendmail"
 	"go.uber.org/zap"
 	"testing"
 )
@@ -34,23 +34,23 @@ func TestServiceSignUp(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessageManager          = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessageManager          = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		props                        incomingProps
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = nil
+	sendmail.MockData.Expected.Error = nil
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
 	languageContent.Data.User.Signup.Email.Request.Subject = append(languageContent.Data.User.Signup.Email.Request.Subject, "")
 	languageContent.Data.User.Signup.Email.Request.Body = append(languageContent.Data.User.Signup.Email.Request.Body, "%s%s")
 
-	cryptManager := cryptographer.NewCryptographer("6368616e676520746869732070617373")
-	jwtManager := &jwt.Token{}
+	cryptManager := secretdata.NewSecretData("6368616e676520746869732070617373")
+	jwtManager := new(jwt.MockDescription)
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
 		DomainAPI:              configDomainAPI,
@@ -68,6 +68,8 @@ func TestServiceSignUp(t *testing.T) {
 		emailMessageManager,
 		authRepo,
 		cryptManager,
+		cryptManager,
+		jwtManager,
 		jwtManager)
 
 	err = newService.SignUpStep1(ctx, logger, model.ServiceSignUpParam{
@@ -93,8 +95,8 @@ func TestServiceConfirmUserEmail_Success(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessage                 = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessage                 = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		message                      string
@@ -102,7 +104,7 @@ func TestServiceConfirmUserEmail_Success(t *testing.T) {
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = nil
+	sendmail.MockData.Expected.Error = nil
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
@@ -110,8 +112,8 @@ func TestServiceConfirmUserEmail_Success(t *testing.T) {
 	languageContent.Data.User.Signup.Email.Password.Body = append(languageContent.Data.User.Signup.Email.Password.Body, "%s%s")
 	languageContent.Data.User.Signup.Page.Text = append(languageContent.Data.User.Signup.Page.Text, "text")
 
-	cryptographerManager := &cryptographer.Cipher{}
-	jwtManager := &jwt.Token{}
+	cryptographerManager := &secretdata.Cipher{}
+	jwtManager := new(jwt.MockDescription)
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
 		DomainAPI:              configDomainAPI,
@@ -126,6 +128,8 @@ func TestServiceConfirmUserEmail_Success(t *testing.T) {
 		emailMessage,
 		authRepo,
 		cryptographerManager,
+		cryptographerManager,
+		jwtManager,
 		jwtManager)
 
 	message, err = newService.SignUpStep2(ctx, logger, "12345")
@@ -152,15 +156,15 @@ func TestServiceConfirmUserEmail_Error(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessage                 = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessage                 = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		contextGetter                = new(middleware.MockDescription)
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = errors.New("REPO_ERROR")
+	sendmail.MockData.Expected.Error = errors.New("REPO_ERROR")
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
@@ -168,8 +172,8 @@ func TestServiceConfirmUserEmail_Error(t *testing.T) {
 	languageContent.Data.User.Signup.Email.Password.Body = append(languageContent.Data.User.Signup.Email.Password.Body, "%s%s")
 	languageContent.Data.User.Signup.Page.Text = append(languageContent.Data.User.Signup.Page.Text, "text")
 
-	cryptographerManager := &cryptographer.Cipher{}
-	jwtManager := &jwt.Token{}
+	cryptographerManager := &secretdata.Cipher{}
+	jwtManager := new(jwt.MockDescription)
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
 		DomainAPI:              configDomainAPI,
@@ -185,6 +189,8 @@ func TestServiceConfirmUserEmail_Error(t *testing.T) {
 		emailMessage,
 		authRepo,
 		cryptographerManager,
+		cryptographerManager,
+		jwtManager,
 		jwtManager)
 
 	_, err = newService.SignUpStep2(ctx, logger, "12345")
@@ -206,8 +212,8 @@ func TestServiceRequestAccessToken(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessage                 = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessage                 = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		device                       model.Device
@@ -215,15 +221,15 @@ func TestServiceRequestAccessToken(t *testing.T) {
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = nil
+	sendmail.MockData.Expected.Error = nil
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
 	languageContent.Data.User.Login.Email.Subject = append(languageContent.Data.User.Login.Email.Subject, "subject")
 	languageContent.Data.User.Login.Email.Body = append(languageContent.Data.User.Login.Email.Body, "%s%s")
 
-	cryptographerManager := &cryptographer.Cipher{}
-	jwtManager := jwt.NewToken("12345", jwt.ParamSigningAlgorithmHS256, "", "", 0, 0)
+	tokenData := &secretdata.Cipher{}
+	token := new(jwt.MockDescription)
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
 		DomainAPI:              configDomainAPI,
@@ -238,17 +244,18 @@ func TestServiceRequestAccessToken(t *testing.T) {
 		emailMessageQueue,
 		emailMessage,
 		authRepo,
-		cryptographerManager,
-		jwtManager)
+		tokenData,
+		tokenData,
+		token,
+		token)
 
-	_, err =
-		newService.CreateAccessToken(ctx, logger, model.ServiceCreateAccessTokenParam{
-			Email:     "email",
-			Password:  "password",
-			ClientID:  "PWA",
-			UserAgent: "userAgent",
-			Device:    device,
-		})
+	_, err = newService.CreateAccessToken(ctx, logger, model.ServiceCreateAccessTokenParam{
+		Email:     "email",
+		Password:  "password",
+		ClientID:  "PWA",
+		UserAgent: "userAgent",
+		Device:    device,
+	})
 
 	if err != nil {
 		t.Errorf("service returned wrong the err value: got %v want %v",
@@ -267,22 +274,22 @@ func TestServiceRefreshAccessToken(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessage                 = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessage                 = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		contextGetter                = new(middleware.MockDescription)
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = nil
+	sendmail.MockData.Expected.Error = nil
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
 	languageContent.Data.User.Login.Email.Subject = append(languageContent.Data.User.Login.Email.Subject, "subject")
 	languageContent.Data.User.Login.Email.Body = append(languageContent.Data.User.Login.Email.Body, "%s%s")
 
-	cryptographerManager := new(cryptographer.MockDescription)
+	cryptographerManager := new(secretdata.MockDescription)
 	jwtManager := new(jwt.MockDescription)
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
@@ -299,6 +306,8 @@ func TestServiceRefreshAccessToken(t *testing.T) {
 		emailMessage,
 		authRepo,
 		cryptographerManager,
+		cryptographerManager,
+		jwtManager,
 		jwtManager)
 
 	_, err = newService.RefreshAccessToken(ctx, logger, "refreshToken")
@@ -320,23 +329,23 @@ func TestServiceRevokeRefreshToken(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessage                 = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessage                 = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		contextGetter                = new(middleware.MockDescription)
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = nil
+	sendmail.MockData.Expected.Error = nil
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
 	languageContent.Data.User.Login.Email.Subject = append(languageContent.Data.User.Login.Email.Subject, "subject")
 	languageContent.Data.User.Login.Email.Body = append(languageContent.Data.User.Login.Email.Body, "%s%s")
 
-	cryptographerManager := new(cryptographer.MockDescription)
-	jwtManager := jwt.NewToken("12345", jwt.ParamSigningAlgorithmHS256, "", "", 0, 0)
+	cryptographerManager := new(secretdata.MockDescription)
+	token := new(jwt.MockDescription)
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
 		DomainAPI:              configDomainAPI,
@@ -352,7 +361,9 @@ func TestServiceRevokeRefreshToken(t *testing.T) {
 		emailMessage,
 		authRepo,
 		cryptographerManager,
-		jwtManager)
+		cryptographerManager,
+		token,
+		token)
 
 	err = newService.RevokeRefreshToken(ctx, logger, model.ServiceRevokeRefreshTokenParam{
 		EncryptedUserData: []byte("encryptedUserData"),
@@ -375,23 +386,23 @@ func TestServiceRequestUserPasswordReset(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessage                 = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessage                 = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		contextGetter                = new(middleware.MockDescription)
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = nil
+	sendmail.MockData.Expected.Error = nil
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
 	languageContent.Data.User.ResetPassword.Email.Request.Subject = append(languageContent.Data.User.Login.Email.Subject, "subject")
 	languageContent.Data.User.ResetPassword.Email.Request.Body = append(languageContent.Data.User.Login.Email.Body, "%s%s")
 
-	cryptographerManager := cryptographer.NewCryptographer("")
-	jwtManager := jwt.NewToken("12345", jwt.ParamSigningAlgorithmHS256, "", "", 0, 0)
+	cryptographerManager := secretdata.NewSecretData("")
+	token := new(jwt.MockDescription)
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
 		DomainAPI:              configDomainAPI,
@@ -407,7 +418,9 @@ func TestServiceRequestUserPasswordReset(t *testing.T) {
 		emailMessage,
 		authRepo,
 		cryptographerManager,
-		jwtManager)
+		cryptographerManager,
+		token,
+		token)
 
 	err = newService.ResetUserPasswordStep1(ctx, logger, "email")
 
@@ -429,23 +442,24 @@ func TestServiceGetListActiveSessions(t *testing.T) {
 		configDomainAPI              = "domain.com"
 		configAuthInviteCodeRequired = true
 		languageContent              config.LanguageContent
-		emailMessageQueue            = make(chan email.MessageBox, 1)
-		emailMessage                 = new(email.AddEmailMessageToQueue_MockDescription)
+		emailMessageQueue            = make(chan sendmail.MessageBox, 1)
+		emailMessage                 = new(sendmail.MockDescription)
 		authRepo                     = new(repository.Mock)
 		err                          error
 		contextGetter                = new(middleware.MockDescription)
 	)
 
 	authRepo.Expected.Error = nil
-	email.AddEmailMessageToQueue_MockData.Expected.Error = nil
+	sendmail.MockData.Expected.Error = nil
 
 	languageContent.Language = make(map[string]int)
 	languageContent.Language["abc"] = 0
 	languageContent.Data.User.Login.Email.Subject = append(languageContent.Data.User.Login.Email.Subject, "subject")
 	languageContent.Data.User.Login.Email.Body = append(languageContent.Data.User.Login.Email.Body, "%s%s")
 
-	cryptographerManager := new(cryptographer.MockDescription)
-	jwtManager := jwt.NewToken("12345", jwt.ParamSigningAlgorithmHS256, "", "", 0, 0)
+	cryptographerManager := new(secretdata.MockDescription)
+	token := new(jwt.MockDescription)
+
 	//goland:noinspection GoBoolExpressions
 	serviceConfig := model.ConfigService{
 		DomainAPI:              configDomainAPI,
@@ -461,7 +475,9 @@ func TestServiceGetListActiveSessions(t *testing.T) {
 		emailMessage,
 		authRepo,
 		cryptographerManager,
-		jwtManager)
+		cryptographerManager,
+		token,
+		token)
 
 	_, err = newService.GetListActiveSessions(ctx, logger, encryptedUserData)