+++ app/information/rest/rest.go
@@ -1,4 +1,4 @@
-/* Copyright © 2020. Financelime, https://financelime.com. All rights reserved.
+/* Copyright © 2021. Financelime, https://financelime.com. All rights reserved.
    Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
    License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
 
@@ -8,7 +8,7 @@ import (
 	"encoding/json"
 	"github.com/dmalix/financelime-authorization/app/information"
 	"github.com/dmalix/financelime-authorization/app/information/model"
-	"github.com/dmalix/financelime-authorization/packages/middleware"
+	"github.com/dmalix/middleware"
 	"go.uber.org/zap"
 	"net/http"
 )