+++ packages/cryptographer/domain.go
@@ -1,6 +1,0 @@
-package cryptographer
-
-type Cryptographer interface {
-	Encrypt(data []byte) ([]byte, error)
-	Decrypt(data []byte) ([]byte, error)
-}