+++ packages/jwt/domain.go
@@ -1,6 +1,0 @@
-package jwt
-
-type JWT interface {
-	GenerateToken(publicSessionID string, privateBox []byte, tokenPurpose string, issuedAt ...int64) (string, error)
-	VerifyToken(jwt string) (JsonWebToken, error)
-}